requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv==1.0.0
requests==2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0
openpyxl>=3.1.0